Remove useless casts to int
In an expression of the form:
[unsigned int] = (int)[size_t];
the cast to int serves only to invoke implementation-defined behavior,
since size_t is unsigned.  Remove such casts from mdc2dgst.c.

Also change the num element of struct mdc2_ctx_st to be a size_t instead
of unsigned int, so as to have the assignments mentioned above store to
a variable of the same width.