Update expectation for dsa-trailingwhitespace.pem
With the pre-refactor PEM parser, trailing whitespace was stripped
prior to any other handling, which masked the line length issue
for this test case.  The printable encoding description in section
4.3.2.4 of RFC 1421 (that imposes these line limits) claims to be
applicable to "PEM message types ENCRYPTED and MIC-ONLY.", so it
seems that other messages are permitted to use more relaxed
encoding rules.  (Note that RFC 1421 predates RFC 2119, so it is
not necessarily clear which normative requirements apply where.)