Defer key updates while writes are pending
If we try to write an application-induced KeyUpdate message while
we are already in the middle of writing application-data records,
we mess up the record framing and cause the connection to fail.
We already deferred peer-requested KeyUpdate messages to a record
boundary, but allowed the application to request a KeyUpdate at
arbitrary times.

Skip setting SSL_in_init() while writes are pending -- we will
check for the key-update state and enter "init" from ssl3_write_bytes()
on the next call.

Fixes: #12485