DO NOT MERGE: work around OSSL_ENCODER issue with EC keys
For some yet-to-be-determined reason, when i2d_provided() attempts to
encode an EVP_PKEY of type EVP_PKEY_EC, the array of potential encoders
to stack up does not include any that will match the default
"type-specific" output format, which AFAICT is supposed to be equivalent
to the "ec" output format for EC keys.  In turn, the encoding fails and
we silently serialize the SSL without the peer's DH key, and the tests
fail because the expected-to-be-present key is absent.

Work around the issue by detecting EVP_PKEY_EC keys in i2d_provided() and
forcing them to use the "ec" output format instead of the non-functional
"type-specific" output format.  A proper fix would be to get the
implementation working with the unchanged "type-specific" output format,
but that fix is not available yet.