Temporary SSL_get_negotiated_group tests for no-DHE resumption
In this working tree not intended for submission to master, we have
modified the behavior of the SSL_OP_ALLOW_NO_DHE_KEX flag (on clients)
to mean that the PSK-only key-exchange mode is the *only* mode sent.
(That is, it is no longer an "allow" flag but an "enforce" flag.)
This enables us to actually negotiate the PSK-only mode between
two openssl endpoints, and thus to be able to test that functionality.
Do so, by adding a PSK-only resumption with different configured groups.
We must get back the original group from the original session, not
anything negotiated on the current connection.